Remove BindingPrecedence and declare <~ with DefaultPrecedence
The standard library includes this curious operator declaration:

    infix operator ~> : DefaultPrecedence

However there are no protocols are free functions that provide any
definition or hint at its intended semantics.

In light of this, it makes a lot of sense for both operators to exist in
the same precedence group. DefaultPrecedence is at a slightly higher
precedence than BindingPrecedence was, now binding more tightly than
the ternary operator. This feels reasonable, as now it's possible to,
e.g., switch between two different binding statements in a ternary
expression.