ViewModelObserver inherits from BindingOwner
Bindings provides the `BindingOwner` protocol, which implicitly extends
all conformers with a `Subscriptions` collection. This makes it
possible, in a generic context, to scope a given binding to the lifetime
of the object that owns the binding.

By inheriting `ViewModelObserver` from `BindingOwner`, all observers
gain the benefit of this automatic collection of subscriptions. As a
side effect, we can update all our examples to remove the explicit
declaration of a `subscriptions` property.