Check if observed object is non-nil before taking send lock
If the observed object has gone out of scope, then we should immediately
send a completion event downstream and finish. This avoids us taking the
lock if there is no longer an object to observe.

Because `finish()` and `cancel()` automatically clean up the upstream
`subscription`, we can simplify the critical section that checks for
pending demand.